# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/parcel-history": {
            "get": {
                "tags": [
                    "parcel-history"
                ],
                "summary": "parcel-history",
                "description": "Returns all parcel-history",
                "operationId": "retrieve_parcel-history",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "parcel"
                ],
                "summary": "submit contact form",
                "description": "add parcel history",
                "operationId": "add parcel history",
                "requestBody": {
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "required": [
                                    "parcel_id",
                                    "tracking_id",
                                    "report"
                                ],
                                "properties": {
                                    "parcel_id": {
                                        "type": "integer"
                                    },
                                    "tracking_id": {
                                        "type": "text"
                                    },
                                    "report": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Create Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Not processed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/parcel-history/{tracking_id}": {
            "get": {
                "tags": [
                    "parcel_details"
                ],
                "summary": "parcel-history-details",
                "description": "parcel-history-details",
                "operationId": "parcel-history-details",
                "parameters": [
                    {
                        "name": "tracking_id",
                        "in": "path",
                        "description": "parcel-history-details",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Not processed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/parcel-history/{id}": {
            "put": {
                "tags": [
                    "update parcel-history"
                ],
                "summary": "update parcel-history",
                "description": "update parcel-history",
                "operationId": "update-parcel-history",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of price",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "parcel_id",
                                    "tracking_id",
                                    "report"
                                ],
                                "properties": {
                                    "parcel_id": {
                                        "type": "integer"
                                    },
                                    "tracking_id": {
                                        "type": "text"
                                    },
                                    "report": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "update Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "update Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Not processed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "delete parcel-history"
                ],
                "summary": "delete parcel-history",
                "description": "delete parcel-history",
                "operationId": "delete-parcel-history",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of parcel-history",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "delete Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "delete Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Not processed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}